+++ src/app/code/community/SchumacherFM/FastIndexer/Helper/Data.php
@@ -14,6 +14,8 @@ class SchumacherFM_FastIndexer_Helper_Data extends Mage_Core_Helper_Abstract
      */
     const CONFIG_DB_NAME = 'global/resources/default_setup/connection/dbname';
 
+    const XPATH_INDEXER_PRODUCT_FLAT_TRANSACTION_DISABLED = 'fastindexer/indexer/disable_transaction_for_product';
+
     /**
      * @var boolean
      */
@@ -279,6 +281,10 @@ class SchumacherFM_FastIndexer_Helper_Data extends Mage_Core_Helper_Abstract
         return
             version_compare(Mage::getVersion(), '1.7.9.9') < 0 &&
             $this->optimizeUrlRewriteFlatCategory();
+    }
+
+    public function isTransactionDisabledForProductFlatReindes(){
+        return Mage::getStoreConfigFlag(self::XPATH_INDEXER_PRODUCT_FLAT_TRANSACTION_DISABLED);
     }
 
     /**