+++ CHANGELOG.md
@@ -7,3 +7,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased 
 - edit validation of db names to give the possibility to use name with underscore
+- remove transaction for catalog_product_flat reindexAll
+- add config to toggle the transaction for catalog_product_flat